requirements.txt
Flask==2.2.2
jsonpatch==1.32
python-dotenv==0.13.0
requests==2.23.0
# mongoengine==0.22.1
boto3==1.26.71
pymysql==1.0.2
flask_sqlalchemy==3.0.2
SQLAlchemy==1.4.45
sqlalchemy_utils==0.36.8
dictalchemy==0.1.1
flask-cors==3.0.10
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2